refactor(gateway): extract handleSessionEvent stream/render dispatcher (#2996 P4-A)
Relocate the ~1,970-line `handleSessionEvent` switch verbatim into
`telegram-plugin/gateway/stream-render.ts` as
`handleSessionEvent(deps, ev)`, leaving gateway.ts with a thin wrapper
(`gatewayStreamRenderDeps()` builder + `StreamRenderDeps` type) that pins
the live singletons and closures.

Why
- gateway.ts is the reliability-defect magnet; P4-A drains ~1,875 net
  lines (30,471 -> 28,593) and ratchets the ceiling down accordingly.

Shared-singleton discipline (plan Amendment 1/5, BLOCKING)
- The answer-stream `outboundDedup` sites (and `backstopDeliveryLedger`)
  use the SAME injected instances P2's `sendReply` receives — never a
  re-`new`. The module constructs neither cache nor ledger.

Turn handle (Amendment 9 / #1664)
- The module never reads the `currentTurn` global; every live re-read
  routes through the injected `getCurrentTurn()` accessor, preserving each
  call site's pin-vs-live choice verbatim. Two other reassigned globals
  (`pendingPtyPartial`, `lastContextExhaustionWarningAt`) route through
  get/set accessors. Deviations from byte-identical are enumerated in the
  module header (the accessor spellings + one capture-once-then-guard for
  TS narrowing across a call expression).

Test plan (Amendment 10 oracle)
- New extracted-module golden harness `stream-render-golden.test.ts`
  drives the REAL `handleSessionEvent` turn-flush path to the REAL
  `outboundDedup.record`, then proves P2's REAL `sendReply` dedups on the
  SAME shared cache (cross-surface, with a second-cache negative control).
  Structural tests pin: exactly one `new OutboundDedupCache` in gateway,
  none in either module; both deps builders inject the same binding; the
  wrapper delegates; the module reads only `getCurrentTurn()`.
- Scoped: 128 tests green across stream/answer/narrative/dedup/turn-flush
  suites; tsc (plugin-references) clean; bot-api-wrapping clean (drift-proof
  inline markers moved with the code — no allowlist edits); ratchet lowered
  to 28,593.

Risk: pure-move refactor, no behavior change. Job spec:
reference/jobs (gateway reliability decomposition, #2996).
