+++ scripts/gateway-line-ratchet.txt
@@ -1,1 +1,1 @@
-30422
+28593
